Fix build by making ControlCheck non-inline
For functions to be inline they apparently
have to be global. Tomorrow, implement
the rest of any predicted motor functions
with the control and then begin refactoring
the code within the scope of the Robot class
provided.

I.E. I was an idiot and glazed over the
Robot::function() in Robot.cpp and thought
First wanted the template to be procedural
(e.g. c-stype) for the sake of simplicity.

Object oriented programming was implemented
in the first place but was glazed over.

OOF