Fix recording utility build breaks
In order to use multithreading a static function or a non-object function must by used in order for  the thing to work out pretty good(although given how compilers work, I don't see the reason they shouldn't). I ended up creating a friend function with a pointer to the Robot Singleton. The I had the Robot call the singleton from itself and from there on, all is well.